% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{add_index}
\alias{add_index}
\title{Create index on a table}
\usage{
add_index(dbconn, tblname, key, idx_name = paste0("idx_",
  digest::digest(tblname)))
}
\arguments{
\item{dbconn}{SQLConnection. A database connection.}

\item{tblname}{character. The name of the table to add an index to.}

\item{key}{character. Identifier of database table.}

\item{idx_name}{character. Name of the index.}
}
\description{
Create index on a table
}

