% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/db.R
\name{build_insert_query}
\alias{build_insert_query}
\title{Build an INSERT query for RPostgreSQL.}
\usage{
build_insert_query(tblname, df)
}
\arguments{
\item{tblname}{character. The table to insert into.}

\item{df}{data.frame. The data to insert.}
}
\value{
a string representing the query that must be executed, to be
   used in conjunction with postgresqlpgExec.
}
\description{
Normally, this would not have to be done manually, but believe it or not,
there appears to be a bug in the built-in dbWriteTable for some complex
data!
}

