% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/utils.R
\name{build_select_by_multiple_keys}
\alias{build_select_by_multiple_keys}
\title{Build a SELECT query for multiple primary keys for RPostgreSQL.}
\usage{
build_select_by_multiple_keys(df, tblname)
}
\arguments{
\item{df}{data.frame. The data.frame of primary keys, with names
corresponding to database column names.}

\item{tblname}{character. The name of the table.}
}
\description{
Turn a data.frame like \code{data.frame(a = c(1, 2, NA), b = (4, NA, 5)}
into the query
 \code{"SELECT * FROM tblname WHERE (a=1 AND b=4) OR (a=2 AND b IS NULL) OR (a is NULL and b=5)"}
}

