% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cache.R
\name{cache}
\alias{cache}
\title{Closure of a cache layer}
\usage{
cache(uncached_function, prefix, key, salt, con)
}
\arguments{
\item{uncached_function}{function. The function to cache.}

\item{prefix}{character. Database table prefix.}

\item{key}{character. Key for the database backend.}

\item{salt}{atomic character vector. Database table salt.}

\item{con}{SQLConnection. Database connection.}
}
\value{
a function with a cache layer of database backend.
}
\description{
Closure of a cache layer
}

