% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{CACHE_METADATA_TABLE}
\alias{CACHE_METADATA_TABLE}
\title{Create cache_meta table}
\format{\preformatted{ chr "cache_metadata"
}}
\usage{
CACHE_METADATA_TABLE
}
\description{
Create cache_meta table
}
\keyword{datasets}

