% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/db.R
\name{cache_salt}
\alias{cache_salt}
\title{Compute the digest of a database salt.}
\usage{
cache_salt(salt)
}
\arguments{
\item{salt}{list. Salt for a cached table, i.e., the list of arguments
passed to the function that necessitate creation of a new table.}
}
\value{
the MD5 digest (32 hexadecimal digit string).
}
\description{
Compute the digest of a database salt.
}

