% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/utils.R
\name{cached_data}
\alias{cached_data}
\title{Fetch cached loan data from local storage.}
\usage{
cached_data(loans, version, root = syberia_root())
}
\arguments{
\item{loans}{numeric. A vector of loan IDs.}

\item{version}{character. The model version for which to fetch data.}

\item{root}{character. Current syberia root project. Default is
\code{syberia_root()}.}
}
\value{
a data.frame of loans that could be fetched from the cache.
}
\description{
Fetch cached loan data from local storage.
}

