% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clear_cache.R
\name{clear_cache}
\alias{clear_cache}
\title{Wipe the cachemeifyoucan caching layer for a cached function.}
\usage{
clear_cache(fn, ...)
}
\arguments{
\item{fn}{function or character.}

\item{...}{additional arguments to the function. If present,
these should specify all the salts of the function. The tables
representing the combination of this \code{fn} and parameters
will be wiped. By default, if only \code{fn} is provided,
wipe \emph{all} tables that cache functions whose prefix
is the \code{fn}.}
}
\value{
\code{TRUE} or \code{FALSE} according as the wipe
  attempt was successful.
}
\description{
Wipe the cachemeifyoucan caching layer for a cached function.
}

