% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{column_names_map}
\alias{column_names_map}
\title{Fetch the map of column names.}
\usage{
column_names_map(dbconn)
}
\arguments{
\item{dbconn}{SQLConnection. A database connection.}
}
\description{
Fetch the map of column names.
}

