% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/db.R
\name{create_table}
\alias{create_shards_table}
\alias{create_table}
\title{Fetch all the shards for the given table name.}
\usage{
create_shards_table(dbconn, tblname)
}
\arguments{
\item{dbconn}{SQLConnection. A database connection.}

\item{tblname}{character.The table to be created}
}
\description{
Fetch all the shards for the given table name.
}

