% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/utils.R
\name{data_sync}
\alias{data_sync}
\title{Sync credit model data between Rails and R}
\usage{
data_sync(version, loans, root = syberia_root())
}
\arguments{
\item{version}{character. The version of the model whose
data we would like to sync.}

\item{loans}{numeric. A list of reference load ids by which
to sync the dataset.}

\item{root}{character. Root of the relevant syberia project.
The default is \code{syberia_root()}. This project must define
a \code{db} configuration option in its syberia config file.
This database will be used for syncing.}
}
\value{
\code{TRUE} or \code{FALSE} according as the sync is
  successful or insuccessful.
}
\description{
Sync credit model data between Rails and R
}

