% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db2df}
\alias{db2df}
\title{Convert the raw fetched database table to a readable data frame.}
\usage{
db2df(df, dbconn, key)
}
\arguments{
\item{df}{dataframe. Raw fetched database table.}

\item{dbconn}{SQLConnection. A database connection.}

\item{key}{character. Identifier of database table.}
}
\description{
Convert the raw fetched database table to a readable data frame.
}

