% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/db.R
\name{db_connection}
\alias{db_connection}
\title{Obtain a connection to a database.}
\usage{
db_connection(database.yml, env = "cache", verbose = TRUE, strict = TRUE)
}
\arguments{
\item{database.yml}{character. The location of the database.yml file
to use. This could, for example, some file in the config directory.}

\item{env}{character. What environment to use. The default is
`"cache"`.}

\item{verbose}{logical. Whether or not to print messages indicating
loading is in progress.}
}
\value{
the database connection.
}
\description{
By default, this function will read from the `cache` environment.
}
\details{
Your database.yml should look like:

development:
  adapter: PostgreSQL
  username: <username>
  password: <password>
  database: <name>
  host: <domain>
  port: <port #>
}

