% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/utils.R
\name{db_for_syberia_project}
\alias{db_for_syberia_project}
\title{Fetch a database connection for a syberia project given
its root.}
\usage{
db_for_syberia_project(root = syberia_root())
}
\arguments{
\item{root}{character. The root of the Syberia project.
The default is \code{syberia_root()}.}
}
\value{
a database connection object.
}
\description{
Fetch a database connection for a syberia project given
its root.
}

