% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{error_fn}
\alias{error_fn}
\title{Stop on given errors and print corresponding error message.}
\usage{
error_fn(data)
}
\arguments{
\item{data}{data.frame.}
}
\description{
Stop on given errors and print corresponding error message.
}

