% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{generate_new_shard_names}
\alias{generate_new_shard_names}
\title{Generate names for new shards}
\usage{
generate_new_shard_names(tbl_name, numshards)
}
\arguments{
\item{tbl_name}{character. The calculated table name for the function.}

\item{numshards}{numeric. The number of shards to generate names for.}
}
\description{
Generate names for new shards
}

