% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_cache_table_salt}
\alias{get_cache_table_salt}
\title{Get the salt used for a cache table}
\usage{
get_cache_table_salt(dbconn, table_names)
}
\arguments{
\item{dbconn}{SQLConnection. A database connection.}

\item{table_names}{character. The table names to inspect.}
}
\description{
Get the salt used for a cache table
}

