% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/utils.R
\name{get_hashed_names}
\alias{get_hashed_names}
\title{MD5 digest of column names.}
\usage{
get_hashed_names(raw_names)
}
\arguments{
\item{raw_names}{character. A character vector of column names.}
}
\value{
the character vector of hashed names.
}
\description{
MD5 digest of column names.
}

