% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_new_key}
\alias{get_new_key}
\title{setdiff current ids with those in the table of the database.}
\usage{
get_new_key(dbconn, tbl_name, ids, key)
}
\arguments{
\item{dbconn}{SQLConnection. The database connection.}

\item{tbl_name}{character. Database table name.}

\item{ids}{vector. A vector of ids.}

\item{key}{character. Identifier of database table.}
}
\description{
setdiff current ids with those in the table of the database.
}

