% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/db.R
\name{get_shards_for_table}
\alias{get_shards_for_table}
\title{Fetch all the shards for the given table name.}
\usage{
get_shards_for_table(dbconn, tbl_name)
}
\arguments{
\item{dbconn}{SQLConnection. A database connection.}

\item{tbl_name}{character. The calculated table name for the function.}
}
\value{
one or many names of the shard tables.
}
\description{
Fetch all the shards for the given table name.
}

