% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/db.R
\name{is_db_connected}
\alias{is_db_connected}
\title{Helper function to check the database connection.}
\usage{
is_db_connected(con)
}
\arguments{
\item{con}{SQLConnection.}
}
\value{
`TRUE` or `FALSE` indicating if the database connection is good.
}
\description{
Helper function to check the database connection.
}

