% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{read_data}
\alias{read_data}
\title{Fetch the data frame from database conditioned on a key.}
\usage{
read_data(dbconn, tblname, ids, key)
}
\arguments{
\item{dbconn}{SQLConnection. A database connection.}

\item{tblname}{character. Database table name.}

\item{ids}{vector. A vector of ids.}

\item{key}{character. Database key.}
}
\description{
Fetch the data frame from database conditioned on a key.
}

