% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/db.R
\name{remove_old_key}
\alias{remove_old_key}
\title{remove old keys to maintain uniqueness of "id" for the sake of force pushing}
\usage{
remove_old_key(dbconn, tbl_name, ids, key)
}
\arguments{
\item{dbconn}{SQLConnection. The database connection.}

\item{tbl_name}{character. Database table name.}

\item{ids}{vector. A vector of ids.}

\item{key}{character. Identifier of database table.}
}
\description{
remove old keys to maintain uniqueness of "id" for the sake of force pushing
}

