% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_rows}
\alias{remove_rows}
\title{Remove rows from a table in the database.}
\usage{
remove_rows(dbconn, tblname, ids, key)
}
\arguments{
\item{dbconn}{SQLConnection. A database connection.}

\item{tblname}{character. Database table name.}

\item{ids}{vector. A vector of ids.}

\item{key}{character. Table key.}
}
\description{
Remove rows from a table in the database.
}

