% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/utils.R
\name{save_loan_data_in_cache}
\alias{save_loan_data_in_cache}
\title{Save loan data into a database cache.}
\usage{
save_loan_data_in_cache(version, uncached_data)
}
\arguments{
\item{version}{character. Model version.}

\item{uncached_data}{dataframe. The data to save.}
}
\description{
Save loan data into a database cache.
}

