% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/migrations.R
\name{table_migrate}
\alias{table_migrate}
\title{Table migrate}
\usage{
table_migrate(cached_fn, table_list)
}
\arguments{
\item{cached_fn}{cached_function.}

\item{table_list}{list. A list on form of
\code{list("old_table_name_1" = "new_table_name_1", ..., "old_table_name_n" = "new_table_name_n")}}
}
\description{
Table migrate
}

