% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/db.R
\name{table_name_}
\alias{table_name_}
\title{Database table name for a given prefix and salt.}
\usage{
table_name_(prefix, salt, register. = FALSE, dbconn = NULL)
}
\arguments{
\item{prefix}{character. Prefix.}

\item{salt}{list. Salt for the table name.}

\item{register.}{logical. Register table metadata.}

\item{dbconn}{SQLConnection. A database connection.}
}
\value{
the table name. This will just be \code{"prefix_"}
  appended with the MD5 hash of the digest of the \code{salt}.
}
\description{
Database table name for a given prefix and salt.
}

