% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metadata.R
\name{track_cache_salt}
\alias{track_cache_salt}
\title{Track cache metadata and register salt if needed.}
\usage{
track_cache_salt(dbconn, table_name, salt)
}
\arguments{
\item{dbconn}{SQLConnection. A database connection.}

\item{table_name}{character.  Name of cache table.}

\item{salt}{list. Actual value of salt before hashing.}
}
\description{
Track cache metadata and register salt if needed.
}

