% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/db.R
\name{translate_column_names}
\alias{translate_column_names}
\title{Translate column names using the column_names table from MD5 to raw.}
\usage{
translate_column_names(names, dbconn)
}
\arguments{
\item{names}{character. A character vector of column names.}

\item{dbconn}{SQLConnection. A database connection.}
}
\description{
Translate column names using the column_names table from MD5 to raw.
}

