% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cache.R
\name{uncached}
\alias{uncached}
\title{Fetch the uncached function}
\usage{
uncached(fn)
}
\arguments{
\item{fn}{function. The function that you want to uncache.}
}
\description{
If applied to a regular function it returns this function
}

