% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cache.R
\name{without_cache}
\alias{without_cache}
\title{Run an expression without any cachemeifyoucan caching layers.}
\usage{
without_cache(expr)
}
\arguments{
\item{expr}{An expression to execute without any caching layers.}
}
\value{
the result of the expression.
}
\description{
Run an expression without any cachemeifyoucan caching layers.
}
\note{
This is not equivalent to the \code{force.} parameter in cached
   functions.
}
\examples{
\dontrun{
without_cache(some_cached_function(1:5))
}
}

