% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{write_data_safely}
\alias{write_data_safely}
\title{Write data.frames to DB addressing pitfalls}
\usage{
write_data_safely(dbconn, tblname, df, key, safe_columns, blacklist)
}
\arguments{
\item{dbconn}{PostgreSQLConnection. The database connection.}

\item{tblname}{character. The table name to write the data into.}

\item{df}{data.frame. The data to write.}

\item{key}{character. The identifier column name.}

\item{safe_columns}{logical. Whether an error should be thrown if the columns returned
are different than what is in the column cache.}

\item{blacklist}{list. A list of values to not cache.}
}
\description{
Write data.frames to DB addressing pitfalls
}

