% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{write_table_shard_map}
\alias{write_table_shard_map}
\title{Register shards in table_shard_map}
\usage{
write_table_shard_map(dbconn, tblname, shard_names)
}
\arguments{
\item{dbconn}{SQLConnection. The database connection.}

\item{tblname}{character. Table name.}

\item{shard_names}{character. Calculated shard names given the table name}
}
\description{
Register shards in table_shard_map
}

