% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/db.R
\name{build_connection}
\alias{build_connection}
\title{Helper function to build the connection.}
\usage{
build_connection(con, env)
}
\arguments{
\item{con}{connection. Could be characters of yaml file path with optional environment,
or a function passed by user to establish the connection, or a database connetion object.}

\item{env}{character. What environment to use when `con` is a the yaml file path.}
}
\value{
a list of database connection and if it can be re-established.
}
\description{
Helper function to build the connection.
}

