% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cache.R
\name{cache}
\alias{cache}
\title{Closure of a cache layer}
\usage{
cache(fn, prefix, salt, key)
}
\arguments{
\item{fn.}{The function to cache.}

\item{prefix.}{Database table prefix.}

\item{salt.}{Database table salt.}

\item{key.}{Keys for the database backend.}
}
\value{
a function with a cache layer of database backend.
}
\description{
Closure of a cache layer
}

