% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{CACHE_METADATA_TABLE}
\alias{CACHE_METADATA_TABLE}
\title{Create cache_meta table}
\format{An object of class \code{character} of length 1.}
\usage{
CACHE_METADATA_TABLE
}
\description{
Create cache_meta table
}
\keyword{datasets}

