% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cachemeifyoucan-package.R
\docType{package}
\name{cachemifyoucan}
\alias{cachemifyoucan}
\alias{cachemifyoucan-package}
\title{One of the most frustrating parts about being a data scientist
is waiting for data or other large downloads. This package offers a caching
layer for arbitrary functions that relies on a PostgreSQL backend.}
\description{
One of the most frustrating parts about being a data scientist
is waiting for data or other large downloads. This package offers a caching
layer for arbitrary functions that relies on a PostgreSQL backend.
}

