% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/db.R
\name{create_shards_table}
\alias{create_shards_table}
\title{Create the table <=> shards map.}
\usage{
create_shards_table(dbconn, tblname)
}
\arguments{
\item{dbconn}{SQLConnection. A database connection.}

\item{tblname}{character.The table to be created}
}
\description{
Create the table <=> shards map.
}

