% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db2df}
\alias{add_index}
\alias{db2df}
\title{Convert the raw fetched database table to a readable data frame.}
\usage{
db2df(df, dbconn, key)

add_index(dbconn, tblname, key, idx_name = paste0("idx_",
  digest::digest(tblname)))
}
\arguments{
\item{dbconn}{SQLConnection. A database connection.}

\item{df.}{Raw fetched database table.}

\item{key.}{Identifier of database table.}

\item{df.}{Raw fetched database table.}

\item{dbconn}{SQLConnection. A database connection.}

\item{key.}{Identifier of database table.}
}
\description{
Convert the raw fetched database table to a readable data frame.

Create index on a table
}

