% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/db.R
\name{db2df}
\alias{db2df}
\title{Convert the raw fetched database table to a readable data frame.}
\usage{
db2df(df, dbconn, key)
}
\arguments{
\item{dbconn}{SQLConnection. A database connection.}

\item{df.}{Raw fetched database table.}

\item{key.}{Identifier of database table.}
}
\description{
Convert the raw fetched database table to a readable data frame.
}

