% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{table_name}
\alias{table_name}
\title{Fetch table name that caches data for a model version.}
\usage{
table_name(version)
}
\arguments{
\item{version}{character. Model version.}
}
\value{
the table name. This will just be \code{"version_"}
  appended with the MD5 hash of the model version.
}
\description{
Fetch table name that caches data for a model version.
}

