% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/db.R
\name{table_name}
\alias{table_name}
\title{Fetch table name that caches data for a model version.}
\usage{
table_name(prefix, version)
}
\arguments{
\item{version}{atomic character vector. Model version.}

\item{previx}{character. Prefix.}
}
\value{
the table name. This will just be \code{"prefix_"}
  appended with the MD5 hash of the model version.
}
\description{
Fetch table name that caches data for a model version.
}

