% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/db.R
\name{table_name}
\alias{table_name}
\title{Database table name for a given prefix and salt.}
\usage{
table_name(prefix, salt)
}
\arguments{
\item{prefix}{character. Prefix.}

\item{salt}{list. Salt for the table name.}
}
\value{
the table name. This will just be \code{"prefix_"}
  appended with the MD5 hash of the digest of the \code{salt}.
}
\description{
Database table name for a given prefix and salt.
}

