% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/migrations.R
\name{update_cache_salt}
\alias{update_cache_salt}
\title{Update the cache when the salt is changed for an existing cache.}
\usage{
update_cache_salt(dbconn, prefix, old_salt, new_salt)
}
\arguments{
\item{dbconn}{SQLConnection. A database connection.}

\item{prefix}{character.}

\item{old_salt}{list. The old salt object.}

\item{new_salt}{list. A new salt object.}
}
\description{
Update the cache when the salt is changed for an existing cache.
}

