% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/db.R
\name{write_data_safely}
\alias{write_data_safely}
\title{Write data.frames to DB addressing pitfalls}
\usage{
write_data_safely(dbconn, tblname, df, key)
}
\arguments{
\item{dbconn}{PostgreSQLConnection. The database connection.}

\item{tblname}{character. The table name to write the data into.}

\item{df}{data.frame. The data to write.}

\item{key}{character. The identifier column name.}
}
\description{
Write data.frames to DB addressing pitfalls
}

