% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/db.R
\name{write_data_safely}
\alias{write_data_safely}
\title{Helper utility for safe IO of a data.frame to a database connection.}
\usage{
write_data_safely(dbconn, tblname, df, key)
}
\arguments{
\item{dbconn}{PostgreSQLConnection. The database connection.}

\item{tblname}{character. The table name to write the data into.}

\item{df}{data.frame. The data to write.}

\item{key}{character. The identifier column name.}
}
\description{
This function will be mindful of two problems: non-existent columns
and long column names.
}
\details{
Since this is meant to be used as a helper function for caching
data, we must take a few precautions. If certain variables are not
available for older data but are introduced for newer data, we
must be careful to create those columns first.

Furthermore, certain column names may be longer than PostgreSQL supports.
To circumvent this problem, this helper function stores an MD5
digest of each column name and maps them using the `column_names`
helper table.

By default, this function assumes any data to be written is not
already present in the table and should be appended. If the table does
not exist, it will be created.
}

